Add files via upload
I stopped on number 14. It all should be working properly and what not. If you have trouble getting the imports to work let me know and I will show you how to add them to your environment variables. At the end of this assignment we will have to convert the .py file to a .ipynb which is no big deal so we can worry about it after we finish.